@@ -1,70 +1,70 @@
 {
-	"name": "angular-pro-app",
-	"version": "0.0.0",
-	"scripts": {
-		"ng": "ng",
-		"start": "ng serve",
-		"build": "ng build",
-		"build:prod": "ng build --prod",
-		"test": "ng test",
-		"lint": "ng lint",
-		"e2e": "ng e2e",
-		"storybook": "start-storybook -p 6006",
-		"build-storybook": "build-storybook"
-	},
-	"private": true,
-	"dependencies": {
-		"@angular/animations": "~8.2.14",
-		"@angular/cdk": "~8.2.3",
-		"@angular/common": "~8.2.14",
-		"@angular/compiler": "~8.2.14",
-		"@angular/core": "~8.2.14",
-		"@angular/fire": "^5.3.0",
-		"@angular/forms": "~8.2.14",
-		"@angular/material": "^8.2.3",
-		"@angular/platform-browser": "~8.2.14",
-		"@angular/platform-browser-dynamic": "~8.2.14",
-		"@angular/router": "~8.2.14",
-		"firebase": "^7.7.0",
-		"hammerjs": "^2.0.8",
-		"rxjs": "~6.5.4",
-		"tslib": "^1.10.0",
-		"zone.js": "~0.10.2"
-	},
-	"devDependencies": {
-		"@angular-devkit/architect": "<0.900 || ^0.900.0-0 || ^9.0.0-0",
-		"@angular-devkit/build-angular": "~0.803.23",
-		"@angular/cli": "~8.3.23",
-		"@angular/compiler-cli": "~8.2.14",
-		"@angular/language-service": "~8.2.14",
-		"@babel/core": "^7.8.3",
-		"@storybook/addon-a11y": "^5.3.9",
-		"@storybook/addon-actions": "^5.3.9",
-		"@storybook/addon-links": "^5.3.9",
-		"@storybook/addon-notes": "^5.3.9",
-		"@storybook/addons": "^5.3.9",
-		"@storybook/angular": "^5.3.9",
-		"@types/jasmine": "~3.5.1",
-		"@types/jasminewd2": "~2.0.8",
-		"@types/node": "~13.5.0",
-		"babel-loader": "^8.0.6",
-		"bootstrap": "^4.4.1",
-		"codelyzer": "^5.2.1",
-		"firebase-tools": "^7.12.1",
-		"fuzzy": "^0.1.3",
-		"inquirer": "^7.0.4",
-		"inquirer-autocomplete-prompt": "^1.0.1",
-		"jasmine-core": "~3.5.0",
-		"jasmine-spec-reporter": "~4.2.1",
-		"karma": "~4.4.1",
-		"karma-chrome-launcher": "~3.1.0",
-		"karma-coverage-istanbul-reporter": "~2.1.1",
-		"karma-jasmine": "~3.1.0",
-		"karma-jasmine-html-reporter": "^1.5.1",
-		"npm-check-updates": "^4.0.1",
-		"protractor": "~5.4.2",
-		"ts-node": "~8.6.2",
-		"tslint": "~6.0.0",
-		"typescript": "~3.5.3"
-	}
+  "name": "angular-pro-app",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "build:prod": "ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~9.0.1",
+    "@angular/cdk": "~8.2.3",
+    "@angular/common": "~9.0.1",
+    "@angular/compiler": "~9.0.1",
+    "@angular/core": "~9.0.1",
+    "@angular/fire": "^5.3.0",
+    "@angular/forms": "~9.0.1",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~9.0.1",
+    "@angular/platform-browser-dynamic": "~9.0.1",
+    "@angular/router": "~9.0.1",
+    "firebase": "^7.7.0",
+    "hammerjs": "^2.0.8",
+    "rxjs": "~6.5.4",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
+  },
+  "devDependencies": {
+    "@angular-devkit/architect": "<0.900 || ^0.900.0-0 || ^9.0.0-0",
+    "@angular-devkit/build-angular": "~0.803.23",
+    "@angular/cli": "~8.3.23",
+    "@angular/compiler-cli": "~9.0.1",
+    "@angular/language-service": "~9.0.1",
+    "@babel/core": "^7.8.3",
+    "@storybook/addon-a11y": "^5.3.9",
+    "@storybook/addon-actions": "^5.3.9",
+    "@storybook/addon-links": "^5.3.9",
+    "@storybook/addon-notes": "^5.3.9",
+    "@storybook/addons": "^5.3.9",
+    "@storybook/angular": "^5.3.9",
+    "@types/jasmine": "~3.5.1",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~13.5.0",
+    "babel-loader": "^8.0.6",
+    "bootstrap": "^4.4.1",
+    "codelyzer": "^5.2.1",
+    "firebase-tools": "^7.12.1",
+    "fuzzy": "^0.1.3",
+    "inquirer": "^7.0.4",
+    "inquirer-autocomplete-prompt": "^1.0.1",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
+    "karma-jasmine": "~3.1.0",
+    "karma-jasmine-html-reporter": "^1.5.1",
+    "npm-check-updates": "^4.0.1",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.6.2",
+    "tslint": "~6.0.0",
+    "typescript": "~3.7.5"
+  }
 }
